backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.0
httpx==0.27.2
openai==1.51.0
pinecone-client==4.1.1
azure-identity==1.17.1
python-dotenv==1.0.1
tenacity==9.0.0
